major refactor of TimeSeries code package
Created a SignalService that is responsible for
- talking to the datasource service to get information about
    - native step
    - backend engine, and
- calculate appropriate down-sampled-step for Influx query
- generate query based on the information above
- pass the query to backend and retrieve TimeSeries object from backend

Release query-building responsibility from the backends. Backends now is
only responsible for light-processing of input query and post-processing
of the returned results.

Make influx backend generic to datasources. It will now alert user if
user tries to use influx backend for unsupported datasources.

Move datasource and backend related calls from the controller to service.

Move relevant code under App\TimeSeries:
- moved App\Expression to App\TimeSeries\Backend\Graphite\Expression
  - Expression is only used for graphite query
- moved App\Humanize to App\TimeSeries\Backend\Graphite\Expression\Humanize
  - Humanize is only used in graphite query expression

Various smaller warning fixes.
