+++ src/Entity/Ioda/DatasourceEntity.php
@@ -42,23 +42,6 @@ class DatasourceEntity
 {
 
     /**
-     * All available down-sample steps an datasource can use.
-     */
-    const ALL_STEPS = [
-        60, 120, 300, 900, 1800, // minute-level [1, 2, 5, 15, 30]
-        3600, 7200, 21600, 43200,  //hour-level [1, 2, 6, 12]
-        86400, 172800, //day-level [1, 2]
-        604800, 1209600, 2419200, //week-level [1, 2, 4]
-        31536000, 63072000, 315360000, //year-level [1, 2, 10]
-    ];
-
-    /**
-     * Array of steps this datasource can down-sample to, including the native step.
-     * @var int[]
-     */
-    private $downsampleSteps;
-
-    /**
      * Constructor
      * @param string $datasource
      * @param string $name
@@ -73,9 +56,7 @@ class DatasourceEntity
         $this->units = $units;
         $this->backend = $backend;
         $this->nativeStep = $nativeStep;
-        $this->downsampleSteps = array_filter($this::ALL_STEPS, function ($v){return $v >= $this->nativeStep;});
     }
-
 
     //////////////////////////
     //////////////////////////
@@ -179,10 +160,6 @@ class DatasourceEntity
     public function getNativeStep(): int
     {
         return $this->nativeStep;
-    }
-
-    public function getDownsampleSteps(): array {
-        return $this->downsampleSteps;
     }
 
     /**
