+++ src/Service/OutagesAlertsService.php
@@ -37,9 +37,7 @@ namespace App\Service;
 
 
 use App\Entity\Outages\OutagesAlert;
-use App\Service\MetadataEntitiesService;
 use App\Repository\OutagesAlertsRepository;
-use CAIDA\Charthouse\WatchtowerBundle\Entity\WatchtowerAlert;
 
 class OutagesAlertsService
 {
