+++ src/TimeSeries/Backend/Graphite/Expression/Functions/PrototypeParameter.php
@@ -33,18 +33,22 @@
  * MAINTENANCE, SUPPORT, UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
  */
 
-namespace App\Expression\Functions;
+namespace App\TimeSeries\Backend\Graphite\Expression\Functions;
 
 use Swagger\Annotations as SWG;
 use Symfony\Component\Serializer\Annotation\Groups;
 
-class PrototypeTag
+class PrototypeParameter
 {
+    public static $TYPE_TIME_SERIES = 'timeSeries';
+    public static $TYPE_STRING = 'string';
+    public static $TYPE_NUMBER = 'number';
+
     /**
      * @Groups({"public"})
      * @SWG\Property(
      *     type="string",
-     *     example="Transform"
+     *     example="Series"
      * )
      */
     private $name;
@@ -53,15 +57,38 @@ class PrototypeTag
      * @Groups({"public"})
      * @SWG\Property(
      *     type="string",
-     *     example="Time series transformation functions"
+     *     example="List of Series (at least 2)"
      * )
      */
     private $description;
 
-    public function __construct($name, $description)
+    /**
+     * @Groups({"public"})
+     * @SWG\Property(
+     *     type="string",
+     *     enum={"timeSeries", "string", "number"},
+     *     example="timeSeries"
+     * )
+     */
+    private $type;
+
+    /**
+     * @Groups({"public"})
+     */
+    private $mandatory;
+
+    /**
+     * @Groups({"public"})
+     */
+    private $multiple;
+
+    public function __construct($name, $description, $type, $mandatory, $multiple)
     {
         $this->setName($name);
         $this->setDescription($description);
+        $this->setType($type);
+        $this->setMandatory($mandatory);
+        $this->setMultiple($multiple);
     }
 
     public function getName(): string
@@ -74,13 +101,44 @@ class PrototypeTag
         $this->name = $name;
     }
 
-    public function getDescription(): ?string
+    public function getDescription(): string
     {
         return $this->description;
     }
 
-    public function setDescription(?string $description): void
+    public function setDescription(string $description): void
     {
         $this->description = $description;
     }
+
+    public function getType(): string
+    {
+        return $this->type;
+    }
+
+    public function setType(string $type): void
+    {
+        $this->type = $type;
+    }
+
+    public function isMandatory(): bool
+    {
+        return $this->mandatory;
+    }
+
+    public function setMandatory(bool $mandatory): void
+    {
+        $this->mandatory = $mandatory;
+    }
+
+    public function isMultiple(): bool
+    {
+        return $this->multiple;
+    }
+
+    public function setMultiple(bool $multiple): void
+    {
+        $this->multiple = $multiple;
+    }
+
 }
